+++ library/oci_preauthenticated_request_facts.py
@@ -151,11 +151,11 @@ def list_preauthenticated_requests(object_storage_client, module):
                 bucket_name,
             )
             optional_list_method_params = ["object_name_prefix"]
-            optional_kwargs = {
-                param: module.params[param]
+            optional_kwargs = dict(
+                (param, module.params[param])
                 for param in optional_list_method_params
                 if module.params.get(param) is not None
-            }
+            )
             existing_preauthenticated_requests = to_dict(
                 oci_utils.list_all_resources(
                     object_storage_client.list_preauthenticated_requests,