+++ library/oci_suppression_facts.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python
-# Copyright (c) 2018, Oracle and/or its affiliates.
+# Copyright (c) 2018, 2019, Oracle and/or its affiliates.
 # This software is made available to you under the terms of the GPL 3.0 license or the Apache 2.0 license.
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 # Apache License v2.0
@@ -164,11 +164,11 @@ def list_suppressions(email_client, module):
                 "time_created_greater_than_or_equal_to",
                 "time_created_less_than",
             ]
-            optional_kwargs = {
-                param: module.params[param]
+            optional_kwargs = dict(
+                (param, module.params[param])
                 for param in optional_list_method_params
                 if module.params.get(param) is not None
-            }
+            )
             existing_suppressions_summary = to_dict(
                 oci_utils.list_all_resources(
                     email_client.list_suppressions,