+++ module_utils/oracle/oci_ce_utils.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2018, Oracle and/or its affiliates.
+# Copyright (c) 2018, 2019, Oracle and/or its affiliates.
 # This software is made available to you under the terms of the GPL 3.0 license or the Apache 2.0 license.
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 # Apache License v2.0
@@ -134,7 +134,7 @@ def delete_and_wait(
                     )
             elif "lifecycle_state" not in resource or resource[
                 "lifecycle_state"
-            ] not in {"DELETING", "DELETED"}:
+            ] not in ["DELETING", "DELETED"]:
                 resource = call_and_wait(
                     client,
                     result,